requirements.txt
playwright
pandas
requests
playwright-stealth
google-generativeai
openpyxl
xlwings
